Move ColorSchemeDropdown to its own file
